chore: lint cleanup, remove unused negative-control CIFs, mean-bin uint16 export
- Remove unused negative-control CIFs (Al_fcc, Fe_bcc, Ni_fcc, TiO2_rutile,
  Ti_hcp_wrong_a) — these were P0 validation controls for the 0617-4d dataset
  and are not needed for the Ti-only 1_R4Q2 workflow
- preprocess_raw.py: replace py4DSTEM's internal bin_R/bin_Q with project-local
  mean-bin that stores uint16 output (avoids float64 bloat + OpenBLAS issues)
- Various minor lint/formatting fixes across pipeline modules
- Update test_workflow.py for current module signatures

+++ src/fourdstem_pipeline/export.py
@@ -329,7 +329,7 @@ def save_profile_png(
         y = int(y1 - frac * (y1 - y0))
         canvas[y, x0:x1 + 1] = 225
 
-    profile = np.asarray(profiles, dtype=np.float32).reshape(-1, profiles.shape[-1]).mean(axis=0)
+    profile = np.asarray(profiles, dtype=np.float32).reshape(-1, profiles.shape[-1]).mean(axis=0, dtype=np.float32)
     radii = np.asarray(radii, dtype=np.float32)
     if radii.size != profile.size:
         radii = np.arange(profile.size, dtype=np.float32)
@@ -1279,7 +1279,7 @@ def mask_center_for_display(
     arr = np.asarray(image, dtype=np.float32).copy()
     h, w = arr.shape
     cy, cx = center_yx if center_yx is not None else ((h - 1) / 2.0, (w - 1) / 2.0)
-    yy, xx = np.indices(arr.shape)
+    yy, xx = np.indices(arr.shape, dtype=np.float32)
     r = np.sqrt((yy - cy) ** 2 + (xx - cx) ** 2)
     if radius_px > 0:
         arr[r < radius_px] = 0.0
@@ -1327,7 +1327,7 @@ def polar_reproject(
     h, w = arr.shape
     cy, cx = center_yx if center_yx is not None else ((h - 1) / 2.0, (w - 1) / 2.0)
 
-    yy, xx = np.indices(arr.shape)
+    yy, xx = np.indices(arr.shape, dtype=np.float32)
     radii = np.sqrt((yy - cy) ** 2 + (xx - cx) ** 2)
     angles = np.arctan2(yy - cy, xx - cx)  # [-π, π]
     angles = np.where(angles < 0, angles + 2 * np.pi, angles)  # [0, 2π]
